pyproject.toml
[project]
name = "bsky-cib"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "atproto>=0.0.58",
    "dash>=2.18.2",
    "dash-cytoscape>=1.0.2",
    "ipykernel>=6.29.5",
    "matplotlib>=3.10.0",
    "nbformat>=5.10.4",
    "networkx>=3.4.2",
    "numpy>=2.2.3",
    "plotly>=6.0.0",
    "python-dotenv>=1.0.1",
    "rich>=13.9.4",
    "pyvis>=0.3.2",
    "scipy>=1.15.2",
    "python-dateutil>=2.9.0.post0",
    "pandas>=2.2.3",
    "dash-table>=5.0.0",
    "uuid>=1.30",
]
